+++ config/v0.2.0/metric_test.go
@@ -6,7 +6,6 @@ package config
 import (
 	"context"
 	"errors"
-	"fmt"
 	"net/url"
 	"reflect"
 	"testing"
@@ -67,7 +66,7 @@ func TestMeterProvider(t *testing.T) {
 							},
 							{
 								Periodic: &PeriodicMetricReader{
-									Exporter: PushMetricExporter{
+									Exporter: MetricExporter{
 										Console: Console{},
 										OTLP:    &OTLPMetric{},
 									},
@@ -123,7 +122,7 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus-no-host",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: PullMetricExporter{
+					Exporter: MetricExporter{
 						Prometheus: &Prometheus{},
 					},
 				},
@@ -134,7 +133,7 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus-no-port",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: PullMetricExporter{
+					Exporter: MetricExporter{
 						Prometheus: &Prometheus{
 							Host: ptr("localhost"),
 						},
@@ -147,10 +146,10 @@ func TestReader(t *testing.T) {
 			name: "pull/prometheus",
 			reader: MetricReader{
 				Pull: &PullMetricReader{
-					Exporter: PullMetricExporter{
+					Exporter: MetricExporter{
 						Prometheus: &Prometheus{
 							Host:              ptr("localhost"),
-							Port:              ptr(8888),
+							Port:              ptr(0),
 							WithoutScopeInfo:  ptr(true),
 							WithoutUnits:      ptr(true),
 							WithoutTypeSuffix: ptr(true),
@@ -168,9 +167,9 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-exporter-invalid-protocol",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol: ptr("http/invalid"),
+							Protocol: "http/invalid",
 						},
 					},
 				},
@@ -181,14 +180,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("http://localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "http://localhost:4318",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -200,14 +199,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-with-path",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("http://localhost:4318/path/123"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "http://localhost:4318/path/123",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -219,13 +218,13 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-no-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
+							Protocol:    "grpc/protobuf",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -237,14 +236,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-exporter-no-scheme",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -256,14 +255,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-invalid-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr(" "),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    " ",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -275,14 +274,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-none-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -294,14 +293,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-delta-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("delta"),
 						},
@@ -314,14 +313,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-cumulative-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("cumulative"),
 						},
@@ -334,14 +333,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-lowmemory-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("lowmemory"),
 						},
@@ -354,14 +353,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-invalid-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("invalid"),
 						},
@@ -374,14 +373,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-grpc-invalid-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("grpc"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "grpc/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("invalid"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -393,14 +392,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("http://localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "http://localhost:4318",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -412,14 +411,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-with-path",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("http://localhost:4318/path/123"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "http://localhost:4318/path/123",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -431,13 +430,13 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-no-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
+							Protocol:    "http/protobuf",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -449,14 +448,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-exporter-no-scheme",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -468,14 +467,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-invalid-endpoint",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr(" "),
+							Protocol:    "http/protobuf",
+							Endpoint:    " ",
 							Compression: ptr("gzip"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -487,14 +486,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-none-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -506,14 +505,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-cumulative-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("cumulative"),
 						},
@@ -526,14 +525,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-lowmemory-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("lowmemory"),
 						},
@@ -546,14 +545,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-delta-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("delta"),
 						},
@@ -566,14 +565,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-invalid-temporality",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("none"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 							TemporalityPreference: ptr("invalid"),
 						},
@@ -586,14 +585,14 @@ func TestReader(t *testing.T) {
 			name: "periodic/otlp-http-invalid-compression",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						OTLP: &OTLPMetric{
-							Protocol:    ptr("http/protobuf"),
-							Endpoint:    ptr("localhost:4318"),
+							Protocol:    "http/protobuf",
+							Endpoint:    "localhost:4318",
 							Compression: ptr("invalid"),
 							Timeout:     ptr(1000),
-							Headers: []NameStringValuePair{
-								{Name: "test", Value: ptr("test1")},
+							Headers: map[string]string{
+								"test": "test1",
 							},
 						},
 					},
@@ -605,7 +604,7 @@ func TestReader(t *testing.T) {
 			name: "periodic/no-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{},
+					Exporter: MetricExporter{},
 				},
 			},
 			wantErr: errors.New("no valid metric exporter"),
@@ -614,7 +613,7 @@ func TestReader(t *testing.T) {
 			name: "periodic/console-exporter",
 			reader: MetricReader{
 				Periodic: &PeriodicMetricReader{
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						Console: Console{},
 					},
 				},
@@ -627,7 +626,7 @@ func TestReader(t *testing.T) {
 				Periodic: &PeriodicMetricReader{
 					Interval: ptr(30_000),
 					Timeout:  ptr(5_000),
-					Exporter: PushMetricExporter{
+					Exporter: MetricExporter{
 						Console: Console{},
 					},
 				},
@@ -880,7 +879,7 @@ func TestView(t *testing.T) {
 				Stream: &ViewStream{
 					Name:          ptr("new_name"),
 					Description:   ptr("new_description"),
-					AttributeKeys: ptr(IncludeExclude{Included: []string{"foo", "bar"}}),
+					AttributeKeys: []string{"foo", "bar"},
 					Aggregation:   &ViewStreamAggregation{Sum: make(ViewStreamAggregationSum)},
 				},
 			},
@@ -1078,104 +1077,35 @@ func TestAggregation(t *testing.T) {
 	}
 }
 
-func TestNewIncludeExcludeFilter(t *testing.T) {
+func TestAttributeFilter(t *testing.T) {
 	testCases := []struct {
 		name          string
-		attributeKeys *IncludeExclude
+		attributeKeys []string
 		wantPass      []string
 		wantFail      []string
 	}{
 		{
 			name:          "empty",
-			attributeKeys: nil,
-			wantPass:      []string{"foo", "bar"},
-			wantFail:      nil,
+			attributeKeys: []string{},
+			wantPass:      nil,
+			wantFail:      []string{"foo", "bar"},
 		},
 		{
-			name: "filter-with-include",
-			attributeKeys: ptr(IncludeExclude{
-				Included: []string{"foo"},
-			}),
-			wantPass: []string{"foo"},
-			wantFail: []string{"bar"},
-		},
-		{
-			name: "filter-with-exclude",
-			attributeKeys: ptr(IncludeExclude{
-				Excluded: []string{"foo"},
-			}),
-			wantPass: []string{"bar"},
-			wantFail: []string{"foo"},
-		},
-		{
-			name: "filter-with-include-and-exclude",
-			attributeKeys: ptr(IncludeExclude{
-				Included: []string{"bar"},
-				Excluded: []string{"foo"},
-			}),
-			wantPass: []string{"bar"},
-			wantFail: []string{"foo"},
+			name:          "filter",
+			attributeKeys: []string{"foo"},
+			wantPass:      []string{"foo"},
+			wantFail:      []string{"bar"},
 		},
 	}
 	for _, tt := range testCases {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := newIncludeExcludeFilter(tt.attributeKeys)
-			require.NoError(t, err)
+			got := attributeFilter(tt.attributeKeys)
 			for _, pass := range tt.wantPass {
 				require.True(t, got(attribute.KeyValue{Key: attribute.Key(pass), Value: attribute.StringValue("")}))
 			}
 			for _, fail := range tt.wantFail {
 				require.False(t, got(attribute.KeyValue{Key: attribute.Key(fail), Value: attribute.StringValue("")}))
 			}
-		})
-	}
-}
-
-func TestNewIncludeExcludeFilterError(t *testing.T) {
-	_, err := newIncludeExcludeFilter(ptr(IncludeExclude{
-		Included: []string{"foo"},
-		Excluded: []string{"foo"},
-	}))
-	require.Equal(t, fmt.Errorf("attribute cannot be in both include and exclude list: foo"), err)
-}
-
-func TestPrometheusReaderOpts(t *testing.T) {
-	testCases := []struct {
-		name        string
-		cfg         Prometheus
-		wantOptions int
-	}{
-		{
-			name:        "no options",
-			cfg:         Prometheus{},
-			wantOptions: 0,
-		},
-		{
-			name: "all set",
-			cfg: Prometheus{
-				WithoutScopeInfo:           ptr(true),
-				WithoutTypeSuffix:          ptr(true),
-				WithoutUnits:               ptr(true),
-				WithResourceConstantLabels: &IncludeExclude{},
-			},
-			wantOptions: 4,
-		},
-		{
-			name: "all set false",
-			cfg: Prometheus{
-				WithoutScopeInfo:           ptr(false),
-				WithoutTypeSuffix:          ptr(false),
-				WithoutUnits:               ptr(false),
-				WithResourceConstantLabels: &IncludeExclude{},
-			},
-			wantOptions: 1,
-		},
-	}
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			opts, err := prometheusReaderOpts(&tt.cfg)
-			require.NoError(t, err)
-			require.Len(t, opts, tt.wantOptions)
 		})
 	}
 }